[windows] move IRSS classes to win32 platform.
