+++ xbmc/input/InputManager.cpp
@@ -48,7 +48,7 @@
 #endif
 
 #ifdef HAS_IRSERVERSUITE
-#include "input/windows/IRServerSuite.h"
+#include "platform/win32/input/IRServerSuite.h"
 #endif
 
 #include "ButtonTranslator.h"
