+++ xbmc/platform/win32/input/CMakeLists.txt
