+++ xbmc/platform/win32/input/IRServerSuite.h
