+++ xbmc/platform/win32/input/IrssMessage.cpp
