+++ xbmc/platform/win32/input/IrssMessage.h
